Move tests to separate modules
