+++ src/forestclaw/geometry.py
@@ -24,11 +24,3 @@ class Patch(pyclaw_geometry.Patch):
         r"""(int) - Block number of current patch, ``default = 0``"""
         self.mpi_rank = 0
         r"""(int) - MPI rank this patch belongs to, ``default = 0``"""
-
-
-def test_forestclaw_patch():
-    """Test the simple extension of the pyclaw.Patch class"""
-
-    patch = Patch(pyclaw.geometry.dimension(0.0, 1.0, 10))
-    patch.block_number = 2
-    patch.mpi_rank = 3
