+++ src/forestclaw/io/test.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+# encoding: utf-8
+r"""
+Test suite for forestclaw.io
+"""
+
+
+def test_forestclaw_input():
+    """Simple test to read in a ForestClaw ASCII file"""
+
+    raise Exception("Test not implemented!")
+
+if __name__ == '__main__':
+    import nose
+    nose.main()
